Fixed upb encoder for field numbers > 2**28.
The encoder was improperly sign-extending the tag to 64 bits.

+++ upb/encode.c
@@ -134,7 +134,8 @@ static void encode_float(upb_encstate *e, float d) {
   encode_fixed32(e, u32);
 }
 
-static void encode_tag(upb_encstate *e, int field_number, int wire_type) {
+static void encode_tag(upb_encstate *e, uint32_t field_number,
+                       uint8_t wire_type) {
   encode_varint(e, (field_number << 3) | wire_type);
 }
